requirements.txt
requests==2.18.4
gunicorn==19.7.1
djangorestframework==3.7.1
django==1.11.6
psycopg2==2.7.3.1
python-novaclient